+++ src/lib/helpers/index.ts
@@ -44,3 +44,9 @@ type Entries<T> = { [K in keyof T]: [K, T[K]] }[keyof T];
 export function ObjectEntries<T extends object>(t: T): Entries<T>[] {
   return Object.entries(t) as any;
 }
+
+export function getRandomEmoji(): "hmm" | "shrug" | "uncertain" {
+  const array = ["hmm", "shrug", "uncertain"] as const
+  const randomIndex = Math.floor(Math.random()*array.length)
+  return array[randomIndex]
+}